+++ packages/orq-rc/src/models/components/publicschedule.ts
@@ -67,7 +67,6 @@ export type PublicSchedule = {
    */
   type: PublicScheduleType;
   updated: Date;
-  workspaceId: string;
 };
 
 /** @internal */
@@ -96,7 +95,6 @@ export const PublicSchedule$inboundSchema: z.ZodType<
   trigger_count: z.number().int(),
   type: PublicScheduleType$inboundSchema,
   updated: z.string().datetime({ offset: true }).transform(v => new Date(v)),
-  workspace_id: z.string(),
 }).transform((v) => {
   return remap$(v, {
     "_id": "id",
@@ -106,7 +104,6 @@ export const PublicSchedule$inboundSchema: z.ZodType<
     "is_active": "isActive",
     "last_triggered_at": "lastTriggeredAt",
     "trigger_count": "triggerCount",
-    "workspace_id": "workspaceId",
   });
 });