+++ packages/orq-rc/RELEASES.md
@@ -8829,3 +8829,13 @@ Based on:
 - [typescript v4.9.0-rc.20] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.20] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.20 - packages/orq-rc
+
+## 2026-05-04 14:36:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.11 (2.881.17) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.21] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.21 - packages/orq-rc